+++ cm/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Breaking: fixed a cyclical dependency introduced in [v0.2.0] between package `encoding/json` and TinyGo `syscall`. To enable JSON support for Component Model `list` and other types, programs must now explicitly import package `cm/json`:
+	`import _ "go.bytecodealliance.org/cm/json"`
+
 ## [v0.2.0] — 2025-03-15
 
 ### Added